+++ client/package.json
@@ -6,19 +6,21 @@
   "scripts": {
     "dev": "vite",
     "build": "run-p type-check \"build-only {@}\" --",
+    "build:prod": "vite build",
     "preview": "vite preview",
+    "serve:prod": "vite preview --port 4173",
     "build-only": "vite build",
     "type-check": "vue-tsc --build --force",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path ../.gitignore",
-    "format": "prettier --write src/"
+    "format": "prettier --write src/",
+    "lh:prod": "lighthouse http://localhost:4173 --output=json --output-path=./lighthouse-prod.json --only-categories=performance,accessibility,best-practices,seo --chrome-flags=\"--headless\""
   },
   "dependencies": {
     "@heroicons/vue": "^2.2.0",
     "axios": "^1.7.3",
     "pinia": "^2.1.7",
     "vue": "^3.4.29",
-    "vue-router": "^4.3.3",
-    "vuetify": "^3.7.9"
+    "vue-router": "^4.3.3"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.8.0",
@@ -44,7 +46,6 @@
     "typescript": "~5.4.0",
     "vite": "^5.3.1",
     "vite-plugin-compression": "^0.5.1",
-    "vite-plugin-vuetify": "^2.0.4",
     "vue-tsc": "^2.2.0"
   }
 }
