+++ client/src/App.vue
@@ -1,23 +1,11 @@
 <script setup lang="ts">
 import { RouterView, useRoute } from 'vue-router'
-import { VApp } from 'vuetify/components'
 import AppNavbar from '@/components/Nav/AppNavbar.vue'
-// import { ref, watch } from 'vue'
 
 const route = useRoute()
-// const showNavbar = ref(false)
-
-// watch(
-//   () => route.name,
-//   () => {
-//     // Cache la navbar lors du changement de page
-//     showNavbar.value = false
-//   }
-// )
-
 </script>
 <template>
-  <v-app class="app">
+  <div class="app">
     <AppNavbar v-show="route.name !== 'welcome'" class="app-navbar" />
     <div id="main-scroll" class="main-scroll" :class="{ 'with-navbar': route.name !== 'welcome' }">
       <RouterView v-slot="{ Component }">
@@ -26,7 +14,7 @@ const route = useRoute()
         </Transition>
       </RouterView>
     </div>
-  </v-app>
+  </div>
 </template>
 <style>
 html,
@@ -52,14 +40,7 @@ body {
   height: 100vh;
   margin: 0;
   padding: 0;
-  background-color: var(--color-background) !important;
-  /* Force the grey background */
-}
-
-/* Reset any default Vuetify margins/padding */
-.v-application {
-  margin: 0 !important;
-  padding: 0 !important;
+  background-color: var(--color-background);
 }
 
 /* Transitions de page smooth */
@@ -91,18 +72,4 @@ body {
 .app-navbar:not([style*="display: none"]) {
   opacity: 1;
 }
-
-/* .navbar-fade-enter-active,
-.navbar-fade-leave-active {
-  transition: opacity 0.5s cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.navbar-fade-enter-from,
-.navbar-fade-leave-to {
-  opacity: 0;
-}
-
-.navbar-fade-enter-active {
-  transition-delay: 0.2s;
-} */
 </style>
