+++ client/src/assets/base.css
@@ -68,17 +68,15 @@ body {
     background-color 0.5s;
   line-height: 1.6;
   font-family:
-    'Source Sans Pro',
+    'Inter',
+    ui-sans-serif,
+    system-ui,
     -apple-system,
     BlinkMacSystemFont,
     'Segoe UI',
     Roboto,
-    Oxygen,
-    Ubuntu,
-    Cantarell,
-    'Fira Sans',
-    'Droid Sans',
     'Helvetica Neue',
+    Arial,
     sans-serif;
   font-size: 15px;
   text-rendering: optimizeLegibility;
@@ -87,5 +85,7 @@ body {
 }
 
 h1 {
-  font-family: 'Rubik Mono One', monospace;
+  font-family: 'Inter', ui-sans-serif, system-ui, sans-serif;
+  font-weight: 900;
+  letter-spacing: 0.05em;
 }
