+++ client/src/assets/fonts.css
@@ -1,0 +1,8 @@
+/* Inter Variable Font - Auto-hébergée pour performance optimale */
+@font-face {
+  font-family: 'Inter';
+  src: url('/fonts/inter/Inter-Variable.woff2') format('woff2');
+  font-weight: 100 900;
+  font-style: normal;
+  font-display: swap;
+}
