+++ client/src/components/UI/SkeletonLoader.vue
