+++ client/src/main.ts
@@ -1,20 +1,16 @@
+import './assets/fonts.css'
 import './assets/main.css'
-import './assets/tailwind.css' // Tailwind CSS (new)
-import '@mdi/font/css/materialdesignicons.css'
-import 'vuetify/styles'
+import './assets/tailwind.css'
 
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 
 import App from './App.vue'
 import router from './router'
-import { createVuetify } from 'vuetify'
 
-const vuetify = createVuetify()
 const app = createApp(App)
 
 app.use(createPinia())
 app.use(router)
-app.use(vuetify)
 
 app.mount('#app')
