+++ client/vite.config.ts
@@ -2,7 +2,6 @@ import { fileURLToPath, URL } from "node:url"
 import { defineConfig } from "vite"
 import vue from "@vitejs/plugin-vue"
 import vueJsx from "@vitejs/plugin-vue-jsx"
-import vuetify from "vite-plugin-vuetify"
 import { visualizer } from "rollup-plugin-visualizer"
 
 // https://vitejs.dev/config/
@@ -23,7 +22,6 @@ export default defineConfig({
   plugins: [
     vue(),
     vueJsx(),
-    vuetify({ autoImport: true }),
     visualizer({
       filename: "dist/stats.html",
       open: false,
